Preflight: replace node buildpack with dockerfile
