+++ test/preflight/fixtures/deploy-node/.dockerignore
@@ -1,0 +1,6 @@
+/.git
+/node_modules
+.dockerignore
+.env
+Dockerfile
+fly.toml
