+++ test/preflight/fixtures/deploy-node/fly.toml
@@ -1,28 +1,31 @@
 app = "{{apps.0}}"
+primary_region = '{{region}}'
 
 [build]
-builder = "heroku/buildpacks:20"
+  dockerfile = 'Dockerfile'
 
 [deploy]
-release_command = "sleep 2"
+  release_command = "sleep 2"
 
 [env]
-TEST_ID = "{{test.id}}"
+  TEST_ID = "{{test.id}}"
 
-[[services]]
-internal_port = 8080
-protocol = "tcp"
+[http_service]
+  internal_port = 8080
+  force_https = true
+  auto_stop_machines = true
+  auto_start_machines = true
+  min_machines_running = 0
+  processes = ['app']
 
-[[services.ports]]
-handlers = ["http"]
-port = 80
+  [[http_service.checks]]
+    grace_period = "5s"
+    interval = "20s"
+    method = "GET"
+    timeout = "5s"
+    path = "/"
 
-[[services.ports]]
-handlers = ["tls", "http"]
-port = 443
-
-[[services.tcp_checks]]
-grace_period = "1s"
-interval = "15s"
-restart_limit = 6
-timeout = "2s"
+[[vm]]
+  memory = '1gb'
+  cpu_kind = 'shared'
+  cpus = 1
