--- test/preflight/fly_deploy_fixtures_test.go
@@ -1,82 +1,0 @@
-//go:build integration
-// +build integration
-
-package preflight
-
-import (
-	"fmt"
-	"path/filepath"
-	"runtime"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	"github.com/superfly/flyctl/test/preflight/testlib"
-)
-
-func getRootPath() string {
-	_, b, _, _ := runtime.Caller(0)
-	return filepath.Dir(b)
-}
-
-func copyFixtureIntoWorkDir(workDir, name string, exclusion []string) error {
-	src := fmt.Sprintf("%s/fixtures/%s", getRootPath(), name)
-	return testlib.CopyDir(src, workDir, exclusion)
-}
-
-func TestFlyDeployBuildpackNodeAppWithRemoteBuilder(t *testing.T) {
-	f := testlib.NewTestEnvFromEnv(t)
-	err := copyFixtureIntoWorkDir(f.WorkDir(), "deploy-node", []string{})
-	require.NoError(t, err)
-
-	flyTomlPath := fmt.Sprintf("%s/fly.toml", f.WorkDir())
-
-	appName := f.CreateRandomAppMachines()
-	require.NotEmpty(t, appName)
-
-	err = testlib.OverwriteConfig(flyTomlPath, map[string]any{
-		"app": appName,
-		"env": map[string]string{
-			"TEST_ID": f.ID(),
-		},
-	})
-	require.NoError(t, err)
-
-	f.Fly("deploy --remote-only --ha=false")
-
-	f.Fly("deploy --remote-only --strategy immediate --ha=false")
-
-	body, err := testlib.RunHealthCheck(fmt.Sprintf("https://%s.fly.dev", appName))
-	require.NoError(t, err)
-
-	require.Contains(t, string(body), fmt.Sprintf("Hello, World! %s", f.ID()))
-}
-
-func TestFlyDeployBasicNodeWithWGEnabled(t *testing.T) {
-	f := testlib.NewTestEnvFromEnv(t)
-	err := copyFixtureIntoWorkDir(f.WorkDir(), "deploy-node", []string{})
-	require.NoError(t, err)
-
-	flyTomlPath := fmt.Sprintf("%s/fly.toml", f.WorkDir())
-
-	appName := f.CreateRandomAppMachines()
-	require.NotEmpty(t, appName)
-
-	err = testlib.OverwriteConfig(flyTomlPath, map[string]any{
-		"app": appName,
-		"env": map[string]string{
-			"TEST_ID": f.ID(),
-		},
-	})
-	require.NoError(t, err)
-
-	f.Fly("wireguard websockets enable")
-
-	f.Fly("deploy --remote-only --ha=false")
-
-	f.Fly("wireguard websockets disable")
-
-	body, err := testlib.RunHealthCheck(fmt.Sprintf("https://%s.fly.dev", appName))
-	require.NoError(t, err)
-
-	require.Contains(t, string(body), fmt.Sprintf("Hello, World! %s", f.ID()))
-}
